+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dapr"
-version = "0.13.0"
+version = "0.14.0"
 authors = ["dapr.io"]
 edition = "2021"
 license-file = "LICENSE"
@@ -33,6 +33,8 @@ axum-test = "12.1.0"
 once_cell = "1.18.0"
 tokio = { version = "1", features = ["full"] }
 uuid = { version = "1.4.0", features = ["v4"] }
+dapr = {path="./"}
+dapr-macros = { path = "macros" }
 
 [[example]]
 name = "client"
