+++ macros/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "dapr-macros"
+version = "0.14.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+async-trait = "0.1"
+log = "0.4"
+axum = "0.6.19"
+syn = {version="2.0.29",features=["full"]}
+quote = "1.0.8"
+dapr = {version="0.14.0", path="../"}
+
+[dev-dependencies]
+
